requirements.txt
Django>=5.2.1
python-dotenv>=1.0.0
whitenoise>=6.6.0
gunicorn>=21.2.0
